Create util for GET, DEL, PATCH for /api/groups
